Move rumEventsCollection/action
