+++ packages/rum-core/src/domain/action/actionCollection.ts
@@ -1,13 +1,13 @@
 import type { ClocksState, Context, Observable } from '@datadog/browser-core'
 import { noop, assign, combine, toServerDuration, generateUUID } from '@datadog/browser-core'
 
-import type { RawRumActionEvent } from '../../../rawRumEvent.types'
-import { ActionType, RumEventType } from '../../../rawRumEvent.types'
-import type { LifeCycle, RawRumEventCollectedData } from '../../lifeCycle'
-import { LifeCycleEventType } from '../../lifeCycle'
-import type { RumConfiguration } from '../../configuration'
-import type { CommonContext } from '../../contexts/commonContext'
-import type { PageStateHistory } from '../../contexts/pageStateHistory'
+import type { RawRumActionEvent } from '../../rawRumEvent.types'
+import { ActionType, RumEventType } from '../../rawRumEvent.types'
+import type { LifeCycle, RawRumEventCollectedData } from '../lifeCycle'
+import { LifeCycleEventType } from '../lifeCycle'
+import type { RumConfiguration } from '../configuration'
+import type { CommonContext } from '../contexts/commonContext'
+import type { PageStateHistory } from '../contexts/pageStateHistory'
 import type { ActionContexts, ClickAction } from './trackClickActions'
 import { trackClickActions } from './trackClickActions'
