+++ packages/rum-core/src/domain/action/clickChain.spec.ts
@@ -1,6 +1,6 @@
 import type { Clock } from '@datadog/browser-core/test'
 import { mockClock } from '@datadog/browser-core/test'
-import { createFakeClick } from '../../../../test'
+import { createFakeClick } from '../../../test'
 import type { ClickChain } from './clickChain'
 import { MAX_DISTANCE_BETWEEN_CLICKS, MAX_DURATION_BETWEEN_CLICKS, createClickChain } from './clickChain'
