+++ packages/rum-core/src/domain/action/clickChain.ts
