+++ packages/rum-core/src/domain/action/getActionNameFromElement.ts
