+++ packages/rum-core/src/domain/action/getSelectorFromElement.ts
