+++ packages/rum-core/src/domain/action/trackClickActions.spec.ts
@@ -1,12 +1,12 @@
 import type { Context, Duration } from '@datadog/browser-core'
 import { addDuration, clocksNow, timeStampNow, relativeNow } from '@datadog/browser-core'
 import { createNewEvent } from '@datadog/browser-core/test'
-import type { TestSetupBuilder } from '../../../../test'
-import { setup, createFakeClick } from '../../../../test'
-import { RumEventType, ActionType, FrustrationType } from '../../../rawRumEvent.types'
-import type { RumEvent } from '../../../rumEvent.types'
-import { LifeCycleEventType } from '../../lifeCycle'
-import { PAGE_ACTIVITY_VALIDATION_DELAY } from '../../waitPageActivityEnd'
+import type { TestSetupBuilder } from '../../../test'
+import { setup, createFakeClick } from '../../../test'
+import { RumEventType, ActionType, FrustrationType } from '../../rawRumEvent.types'
+import type { RumEvent } from '../../rumEvent.types'
+import { LifeCycleEventType } from '../lifeCycle'
+import { PAGE_ACTIVITY_VALIDATION_DELAY } from '../waitPageActivityEnd'
 import type { ActionContexts } from './actionCollection'
 import type { ClickAction } from './trackClickActions'
 import { finalizeClicks, CLICK_ACTION_MAX_DURATION, trackClickActions } from './trackClickActions'
