+++ packages/rum-core/src/domain/action/trackClickActions.ts
@@ -12,13 +12,13 @@ import {
   ONE_SECOND,
   elapsed,
 } from '@datadog/browser-core'
-import type { FrustrationType } from '../../../rawRumEvent.types'
-import { ActionType } from '../../../rawRumEvent.types'
-import type { RumConfiguration } from '../../configuration'
-import type { LifeCycle } from '../../lifeCycle'
-import { LifeCycleEventType } from '../../lifeCycle'
-import { trackEventCounts } from '../../trackEventCounts'
-import { PAGE_ACTIVITY_VALIDATION_DELAY, waitPageActivityEnd } from '../../waitPageActivityEnd'
+import type { FrustrationType } from '../../rawRumEvent.types'
+import { ActionType } from '../../rawRumEvent.types'
+import type { RumConfiguration } from '../configuration'
+import type { LifeCycle } from '../lifeCycle'
+import { LifeCycleEventType } from '../lifeCycle'
+import { trackEventCounts } from '../trackEventCounts'
+import { PAGE_ACTIVITY_VALIDATION_DELAY, waitPageActivityEnd } from '../waitPageActivityEnd'
 import type { ClickChain } from './clickChain'
 import { createClickChain } from './clickChain'
 import { getActionNameFromElement } from './getActionNameFromElement'
