+++ packages/rum-core/src/domain/lifeCycle.ts
@@ -6,7 +6,7 @@ import type { RawRumEvent } from '../rawRumEvent.types'
 import type { RumEvent } from '../rumEvent.types'
 import type { CommonContext } from './contexts/commonContext'
 import type { RequestCompleteEvent, RequestStartEvent } from './requestCollection'
-import type { AutoAction } from './rumEventsCollection/action/actionCollection'
+import type { AutoAction } from './action/actionCollection'
 import type { ViewEvent, ViewCreatedEvent, ViewEndedEvent } from './rumEventsCollection/view/trackViews'
 
 export const enum LifeCycleEventType {
