+++ packages/rum-core/src/index.ts
@@ -28,7 +28,7 @@ export { getMutationObserverConstructor } from './browser/domMutationObservable'
 export { initViewportObservable, getViewportDimension } from './browser/viewportObservable'
 export { getScrollX, getScrollY } from './browser/scroll'
 export { RumInitConfiguration, RumConfiguration } from './domain/configuration'
-export { DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from './domain/rumEventsCollection/action/getActionNameFromElement'
-export { STABLE_ATTRIBUTES } from './domain/rumEventsCollection/action/getSelectorFromElement'
+export { DEFAULT_PROGRAMMATIC_ACTION_NAME_ATTRIBUTE } from './domain/action/getActionNameFromElement'
+export { STABLE_ATTRIBUTES } from './domain/action/getSelectorFromElement'
 export * from './browser/htmlDomUtils'
 export { getSessionReplayUrl } from './domain/getSessionReplayUrl'
