+++ packages/rum-core/test/testSetupBuilder.ts
@@ -11,7 +11,7 @@ import type { UrlContexts } from '../src/domain/contexts/urlContexts'
 import type { ViewContexts } from '../src/domain/contexts/viewContexts'
 import type { RawRumEventCollectedData } from '../src/domain/lifeCycle'
 import { LifeCycle, LifeCycleEventType } from '../src/domain/lifeCycle'
-import type { ActionContexts } from '../src/domain/rumEventsCollection/action/actionCollection'
+import type { ActionContexts } from '../src/domain/action/actionCollection'
 import type { RumSessionManager } from '../src/domain/rumSessionManager'
 import { RumSessionPlan } from '../src/domain/rumSessionManager'
 import type { RawRumEvent, RumContext } from '../src/rawRumEvent.types'
